% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_sub.R
\name{facet_sub}
\alias{facet_sub}
\title{facet_sub to produce a sub panel around main panel}
\usage{
facet_sub(
  position = c("left", "right", "top", "bottom"),
  shrink = TRUE,
  addGrob = grid::gList(rectGrob()),
  stripTitle = c("Main panel", "Sub panel"),
  panelWidth = c(1, 1)
)
}
\arguments{
\item{position}{the position of sub panel, one of the "left","right","top" and "bottom".}

\item{shrink}{shrink}

\item{addGrob}{the grobs to be added into sub panel, default grid::gList(rectGrob()).}

\item{stripTitle}{the strip title, default "Main panel" and "Sub panel".}

\item{panelWidth}{the relative width of main and sub panel, default c(1,1).}
}
\description{
facet_sub to produce a sub panel around main panel
}
