% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint_motif_vis.R
\name{footprintPlot}
\alias{footprintPlot}
\title{plot footprint line plot from rgt-hint output}
\usage{
footprintPlot(linePath = NULL, TFname = NULL)
}
\arguments{
\item{linePath}{the rgt-hint output directory, default NULL.}

\item{TFname}{the transcription factor name to be plotted, default NULL.}
}
\value{
a ggplot
}
\description{
plot footprint line plot from rgt-hint output
}
