% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint_motif_vis.R
\name{getSeqFrombed}
\alias{getSeqFrombed}
\title{extract sequence from bed}
\usage{
getSeqFrombed(mbpsPath = NULL, TFname = NULL, genome = NULL)
}
\arguments{
\item{mbpsPath}{the mbps file from rgt-motifanalysis output directory, default NULL.}

\item{TFname}{the transcription factor name to be plotted, default NULL.}

\item{genome}{the BSgenome object for genome sequence, default NULL.}
}
\value{
sequence
}
\description{
extract sequence from bed
}
