% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\name{gglogo,homerResult-method}
\alias{gglogo,homerResult-method}
\title{method for gglogo}
\usage{
\S4method{gglogo}{homerResult}(object, type = c("known", "novo"), ...)
}
\arguments{
\item{object}{homerResult.}
}
\value{
ggplot object
}
\description{
method for gglogo
}
