% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\name{gglogo}
\alias{gglogo}
\title{motif logo plot}
\usage{
gglogo(object, type = c("known", "novo"), ...)
}
\arguments{
\item{object}{homerResult.}

\item{type}{the homer output type results to choose, "known" or "novo".}

\item{...}{other args can be passed with ggseqLogo.}
}
\description{
motif logo plot
}
