% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{homerResult-class}
\alias{homerResult-class}
\alias{homerResult}
\title{create homerResult object class}
\value{
homerResult object
}
\description{
create homerResult object class
}
\section{Slots}{

\describe{
\item{\code{known}}{res data.frame. known meta info.}

\item{\code{known}}{motif PWM matrix list. motif PWM matrix.}

\item{\code{known}}{motif PFM matrix list. motif PFM matrix.}

\item{\code{novo}}{res data.frame. novo meta info.}

\item{\code{novo}}{motif PWM matrix list. motif PWM matrix.}

\item{\code{novo}}{motif PFM matrix list. motif PFM matrix.}
}}

