% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseHomer.R
\name{loadHomerRes}
\alias{loadHomerRes}
\title{parse homer output and assemble into homerResult object}
\usage{
loadHomerRes(homerDir = NULL, motifIndex = NULL)
}
\arguments{
\item{homerDir}{the path of homer output data.}

\item{motifIndex}{the numbers of motif index.}
}
\value{
homerResult object
}
\description{
parse homer output and assemble into homerResult object
}
