% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseHomer.R
\name{parseHomer}
\alias{parseHomer}
\title{parse homer output into meta information}
\usage{
parseHomer(homerDir = NULL, motifIndex = NULL, type = c("known", "novo"))
}
\arguments{
\item{homerDir}{the path of homer output data.}

\item{motifIndex}{the numbers of motif index.}

\item{type}{the homer output type results to choose, "known" or "novo".}
}
\value{
data.frame
}
\description{
parse homer output into meta information
}
