% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseHomer.R
\name{parseHomerMotif}
\alias{parseHomerMotif}
\title{extract pwm matrix from motif}
\usage{
parseHomerMotif(homerDir = NULL, motifIndex = NULL, type = c("known", "novo"))
}
\arguments{
\item{homerDir}{the path of homer output data.}

\item{motifIndex}{the numbers of motif index.}

\item{type}{the homer output type results to choose, "known" or "novo".}
}
\value{
pwmList
}
\description{
extract pwm matrix from motif
}
