% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\name{plotMotifHeatmap,homerResult-method}
\alias{plotMotifHeatmap,homerResult-method}
\title{plotMotifHeatmap method}
\usage{
\S4method{plotMotifHeatmap}{homerResult}(
  object,
  type = c("known", "novo"),
  ic.scale = TRUE,
  ymax = 2,
  clustForMotif = FALSE,
  cluster_rows = TRUE,
  logoSpace = 0.5,
  logoWidth = 0.8,
  rightAnnoWidth = 0.5,
  barWidth = 3,
  col1 = NULL,
  col2 = NULL,
  ...
)
}
\arguments{
\item{object}{homerResult.}
}
\description{
plotMotifHeatmap method
}
