% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\name{plotMotifHeatmap}
\alias{plotMotifHeatmap}
\title{heatmap plot for homerResult}
\usage{
plotMotifHeatmap(
  object,
  type = c("known", "novo"),
  ic.scale = TRUE,
  ymax = 2,
  clustForMotif = FALSE,
  cluster_rows = TRUE,
  logoSpace = 0.5,
  logoWidth = 0.8,
  rightAnnoWidth = 0.5,
  barWidth = 3,
  col1 = NULL,
  col2 = NULL,
  ...
)
}
\arguments{
\item{object}{homerResult.}

\item{type}{the homer output type results to choose, "known" or "novo".}

\item{ic.scale}{whether scale to bits or probability, default TRUE.}

\item{ymax}{seqlogo height, default 2.}

\item{clustForMotif}{whether cluster rows by motif similarity, default FALSE.}

\item{cluster_rows}{whether cluster rows, default TRUE.}

\item{logoSpace}{the space between motif logo and heatmap, default 0.5.}

\item{logoWidth}{motif logo width, default 0.8.}

\item{rightAnnoWidth}{right annotation width, default 0.5.}

\item{barWidth}{the match score for de novo results annotation width, default 3.}

\item{col1}{the color for negLogPvalue heatmap, for example: col1 = colorRamp2(c(-2, 0, 2), c("green", "white", "red")).}

\item{col2}{the color for log2Enrichment heatmap.}

\item{...}{other args can be passed with ComplexHeatmap.}
}
\value{
Complexheatmap list
}
\description{
heatmap plot for homerResult
}
