% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseHomer.R
\name{preparePFMmat}
\alias{preparePFMmat}
\title{extract pfm matrix from motif}
\usage{
preparePFMmat(homerDir = NULL, motifIndex = NULL, type = c("known", "novo"))
}
\arguments{
\item{homerDir}{the path of homer output data.}

\item{motifIndex}{the numbers of motif index.}

\item{type}{the homer output type results to choose, "known" or "novo".}
}
\value{
pfmList
}
\description{
extract pfm matrix from motif
}
