% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint_motif_vis.R
\name{seq2pwmMatrix}
\alias{seq2pwmMatrix}
\title{create a pwmMatrix}
\usage{
seq2pwmMatrix(seqs = NULL, shift = c(-10, 10))
}
\arguments{
\item{seqs}{sequence to make pwmMatrix, default NULL.}

\item{shift}{a number of shift to zoom the motif, default c(-10,10).}
}
\value{
A list pwmMatrix
}
\description{
create a pwmMatrix
}
