% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqLogoGrob2.R
\name{seqLogoGrob2}
\alias{seqLogoGrob2}
\title{Create a simple sequence logo grob(slightly modified with seqLogoGrob).}
\usage{
seqLogoGrob2(
  x,
  xmax = NULL,
  ymax = 2,
  ic.scale = TRUE,
  xjust = c("left", "center", "right"),
  vp = NULL
)
}
\arguments{
\item{x}{A \code{\link[TFBSTools]{PFMatrix}} object}

\item{xmax}{A numeric scalar with the maximal width for the logo
(in base-pairs). A value of \code{NULL} will scale the logo to the
full width of the viewport.}

\item{ymax}{A numeric scalar with the maximal height for the logo (in bits)
A value of \code{NULL} will scale the logo to the full height of the
viewport.}

\item{ic.scale}{whether scale to bits or probability, default TRUE.}

\item{xjust}{A character scalar specifying the horizontal adjustment of the
sequence log withint the viewport; one of \code{"left"}, \code{"center"}
or \code{"right"}.}

\item{vp}{vp for polygonGrob, default null.}
}
\value{
A polygon grob.
}
\description{
Create a simple sequence logo grob (grid-graphics object) for a
transcription factor from a position frequency matrix. The logo drawing
code is a simplified version from \code{\link[seqLogo]{seqLogo}} and for
example can be used to embedd sequence logos within other plots.
}
