% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\name{show,homerResult-method}
\alias{show,homerResult-method}
\title{Default method}
\usage{
\S4method{show}{homerResult}(object)
}
\arguments{
\item{object}{homerResult.}
}
\description{
Default method
}
