% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_motif_grob.R
\name{zoom_motif_grob}
\alias{zoom_motif_grob}
\title{construct zoom_motif_grob}
\usage{
zoom_motif_grob(
  pwmMatrix = NULL,
  rel_height = c(0.4, 0.2, 0.4),
  link_gp_col = "white",
  link_gp_fill = "grey90",
  border_col = "white",
  name = NULL,
  gp = NULL,
  vp = NULL
)
}
\arguments{
\item{pwmMatrix}{a pwmMatrix from seq2pwmMatrix function, default NULL.}

\item{rel_height}{the relative height for each part, default c(0.4,0.2,0.4).}

\item{link_gp_col}{the color of link, default "white".}

\item{link_gp_fill}{the fill of link, default "grey90".}

\item{border_col}{the border color, default "white".}

\item{name}{name}

\item{gp}{gp}

\item{vp}{vp}
}
\value{
a grob
}
\description{
construct zoom_motif_grob
}
